requirements.txt
python-telegram-bot==20.3
noisereduce==3.0.0
librosa==0.10.0
soundfile==0.12.1
numpy==1.24.3
scipy==1.10.1
pydub==0.25.1